Project ready to present demo. Details left

+++ components/PaginaPedidos/SeguimientoPedido.jsx
@@ -3,40 +3,42 @@ import { Text, View, StyleSheet } from 'react-native';
 import useDeleteAllFoodOnCart from '../../hooks/useDeleteAllCartOnFood';
 
 const messages = [
-    'El lugar de comidas está preparando tu pedido.',
-    'El delivery llegó al lugar de comidas.',
-    'Tu pedido está listo.',
-    'El delivery tiene tu pedido.',
-    'Tu pedido está en camino.',
-    'Tu pedido fue entregado. Muchas gracias por usar Gula.'
+    '20:05: El lugar de comidas está preparando tu pedido.',
+    '20:15: El delivery llegó al lugar de comidas.',
+    '20:30: Tu pedido está listo.',
+    '20:35: El delivery tiene tu pedido.',
+    '20:40: Tu pedido está en camino.',
+    '20:50: Tu pedido fue entregado.',
+    'Muchas gracias por usar Gula.'
 ];
 
 export default function SeguimientoPedido({ onCompletion }) {
     const [index, setIndex] = useState(0);
-    const [currentMessage, setCurrentMessage] = useState(messages[index]);
+    const [currentMessage, setCurrentMessage] = useState(messages[0]);
     const { deleteAllFoodOnCart, deleteAllFoodOnCartLoading, deleteAllFoodOnCartError } = useDeleteAllFoodOnCart();
 
     useEffect(() => {
-        const interval = setInterval(() => {
-            setIndex((prevIndex) => (prevIndex + 1) % messages.length); // Avanza al siguiente mensaje circularmente
-        }, 3000);
+        if (index < messages.length - 1) {
+            const interval = setInterval(() => {
+                setIndex((prevIndex) => prevIndex + 1);
+            }, 3000);
 
-        // Limpiar el carrito y refetch cuando se muestran todos los mensajes
-        setTimeout(async () => {
-            try {
-                await deleteAllFoodOnCart();
-                console.log("Success", "All food items deleted successfully");
-                onCompletion(); // Llama a la función de callback para indicar que se ha completado el proceso
-            } catch (err) {
-                console.log("Error", "Failed to delete all food items on cart");
-            }
-        }, messages.length * 3000); // 3000 ms por mensaje
-
-        return () => clearInterval(interval); // Limpia el intervalo cuando el componente se desmonta
-    }, []);
+            return () => clearInterval(interval);
+        } else {
+            setTimeout(async () => {
+                try {
+                    await deleteAllFoodOnCart();
+                    console.log("Success", "All food items deleted successfully");
+                    onCompletion();
+                } catch (err) {
+                    console.log("Error", "Failed to delete all food items on cart");
+                }
+            }, 3000); // Delay to allow the last message to be shown for 3 seconds
+        }
+    }, [index]);
 
     useEffect(() => {
-        setCurrentMessage(messages[index]); // Actualiza el mensaje actual cada vez que cambia el índice
+        setCurrentMessage(messages[index]);
     }, [index]);
 
     return (
@@ -49,14 +51,15 @@ export default function SeguimientoPedido({ onCompletion }) {
 const styles = StyleSheet.create({
     textContainer: {
         width: '90%',
-        height: 100,
-        //backgroundColor: 'pink',
+        minHeight: 100,        
         marginTop: 20,
         alignItems: 'center',
         justifyContent: 'center',
         borderRadius: 10,
         borderWidth: 2,
-        borderColor: 'red'        
+        borderColor: 'red',
+        paddingVertical: 20,
+        paddingHorizontal: 20       
     },
     textSeguimiento: {
         fontSize: 16,
